PreCourse_1 Exercise_3 Renamed to LinkedList.java
